+++ src/dye/agents.py
@@ -42,7 +42,7 @@ class AgentBase(abc.ABC):
     class name, ie:
 
     EnvironmentVariables -> environment_variables
-    LsColors -> ls_colors
+    GnuLs -> gnu_ls
     """
 
     classmap = {}
@@ -429,6 +429,96 @@ class Fzf(AgentBase):
         return attribs
 
 
+class GnuLs(AgentBase, LsColorsFromStyle):
+    "Create LS_COLORS environment variable for use with GNU ls"
+
+    LS_COLORS_BASE_MAP = {
+        # map both a friendly name and the "real" name
+        "text": "no",
+        "file": "fi",
+        "directory": "di",
+        "symlink": "ln",
+        "multi_hard_link": "mh",
+        "pipe": "pi",
+        "socket": "so",
+        "door": "do",
+        "block_device": "bd",
+        "character_device": "cd",
+        "broken_symlink": "or",
+        "missing_symlink_target": "mi",
+        "setuid": "su",
+        "setgid": "sg",
+        "sticky": "st",
+        "other_writable": "ow",
+        "sticky_other_writable": "tw",
+        "executable_file": "ex",
+        "file_with_capability": "ca",
+    }
+    # this map allows you to either use the 'native' color code, or the
+    # 'friendly' name defined by shell-themer
+    LS_COLORS_MAP = {}
+    for friendly, actual in LS_COLORS_BASE_MAP.items():
+        LS_COLORS_MAP[friendly] = actual
+        LS_COLORS_MAP[actual] = actual
+
+    def run(self):
+        "Render a LS_COLORS variable suitable for GNU ls"
+        outlist = []
+        havecodes = []
+        # figure out if we are clearing builtin styles
+        try:
+            clear_builtin = self.scopedef["clear_builtin"]
+            if not isinstance(clear_builtin, bool):
+                raise DyeSyntaxError(
+                    f"scope '{self.scope}' requires 'clear_builtin' to be true or false"
+                )
+        except KeyError:
+            clear_builtin = False
+
+        # iterate over the styles given in our configuration
+        for name, style in self.scope_styles.items():
+            if style:
+                mapcode, render = self.ls_colors_from_style(
+                    name,
+                    style,
+                    self.LS_COLORS_MAP,
+                    allow_unknown=False,
+                    scope=self.scope,
+                )
+                havecodes.append(mapcode)
+                outlist.append(render)
+
+        if clear_builtin:
+            style = rich.style.Style.parse("default")
+            # go through all the color codes, and render them with the
+            # 'default' style and add them to the output
+            for name, code in self.LS_COLORS_BASE_MAP.items():
+                if code not in havecodes:
+                    _, render = self.ls_colors_from_style(
+                        name,
+                        style,
+                        self.LS_COLORS_MAP,
+                        allow_unknown=False,
+                        scope=self.scope,
+                    )
+                    outlist.append(render)
+
+        # process the filesets
+
+        # figure out which environment variable to put it in
+        try:
+            varname = self.scopedef["environment_variable"]
+            varname = self.jinja_env.from_string(varname).render()
+        except KeyError:
+            varname = "LS_COLORS"
+
+        # even if outlist is empty, we have to set the variable, because
+        # when we are switching a theme, there may be contents in the
+        # environment variable already, and we need to tromp over them
+        # we chose to set the variable to empty instead of unsetting it
+        return f'''export {varname}="{':'.join(outlist)}"'''
+
+
 class Iterm(AgentBase):
     "Send escape sequences to iTerm terminal emulator"
 
@@ -548,96 +638,6 @@ class Iterm(AgentBase):
         except KeyError:
             # the given style doesn't exist
             pass
-
-
-class LsColors(AgentBase, LsColorsFromStyle):
-    "Create LS_COLORS environment variable for use with GNU ls"
-
-    LS_COLORS_BASE_MAP = {
-        # map both a friendly name and the "real" name
-        "text": "no",
-        "file": "fi",
-        "directory": "di",
-        "symlink": "ln",
-        "multi_hard_link": "mh",
-        "pipe": "pi",
-        "socket": "so",
-        "door": "do",
-        "block_device": "bd",
-        "character_device": "cd",
-        "broken_symlink": "or",
-        "missing_symlink_target": "mi",
-        "setuid": "su",
-        "setgid": "sg",
-        "sticky": "st",
-        "other_writable": "ow",
-        "sticky_other_writable": "tw",
-        "executable_file": "ex",
-        "file_with_capability": "ca",
-    }
-    # this map allows you to either use the 'native' color code, or the
-    # 'friendly' name defined by shell-themer
-    LS_COLORS_MAP = {}
-    for friendly, actual in LS_COLORS_BASE_MAP.items():
-        LS_COLORS_MAP[friendly] = actual
-        LS_COLORS_MAP[actual] = actual
-
-    def run(self):
-        "Render a LS_COLORS variable suitable for GNU ls"
-        outlist = []
-        havecodes = []
-        # figure out if we are clearing builtin styles
-        try:
-            clear_builtin = self.scopedef["clear_builtin"]
-            if not isinstance(clear_builtin, bool):
-                raise DyeSyntaxError(
-                    f"scope '{self.scope}' requires 'clear_builtin' to be true or false"
-                )
-        except KeyError:
-            clear_builtin = False
-
-        # iterate over the styles given in our configuration
-        for name, style in self.scope_styles.items():
-            if style:
-                mapcode, render = self.ls_colors_from_style(
-                    name,
-                    style,
-                    self.LS_COLORS_MAP,
-                    allow_unknown=False,
-                    scope=self.scope,
-                )
-                havecodes.append(mapcode)
-                outlist.append(render)
-
-        if clear_builtin:
-            style = rich.style.Style.parse("default")
-            # go through all the color codes, and render them with the
-            # 'default' style and add them to the output
-            for name, code in self.LS_COLORS_BASE_MAP.items():
-                if code not in havecodes:
-                    _, render = self.ls_colors_from_style(
-                        name,
-                        style,
-                        self.LS_COLORS_MAP,
-                        allow_unknown=False,
-                        scope=self.scope,
-                    )
-                    outlist.append(render)
-
-        # process the filesets
-
-        # figure out which environment variable to put it in
-        try:
-            varname = self.scopedef["environment_variable"]
-            varname = self.jinja_env.from_string(varname).render()
-        except KeyError:
-            varname = "LS_COLORS"
-
-        # even if outlist is empty, we have to set the variable, because
-        # when we are switching a theme, there may be contents in the
-        # environment variable already, and we need to tromp over them
-        # we chose to set the variable to empty instead of unsetting it
-        return f'''export {varname}="{':'.join(outlist)}"'''
 
 
 class Shell(AgentBase):